+++ pkg/component/observability/logging/fluentoperator/crds.go
@@ -7,16 +7,13 @@ package fluentoperator
 import (
 	"context"
 	_ "embed"
-	"time"
 
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/gardener/gardener/pkg/client/kubernetes"
 	"github.com/gardener/gardener/pkg/component"
 	"github.com/gardener/gardener/pkg/utils/flow"
-	"github.com/gardener/gardener/pkg/utils/kubernetes/health"
-	"github.com/gardener/gardener/pkg/utils/retry"
+	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 )
 
 var (
@@ -109,42 +106,9 @@ func (c *crds) Destroy(ctx context.Context) error {
 	return flow.Parallel(fns...)(ctx)
 }
 
-var (
-	// IntervalWaitForCRD is the interval used while waiting for the CRDs to become healthy
-	// or deleted.
-	IntervalWaitForCRD = 1 * time.Second
-	// TimeoutWaitForCRD is the timeout used while waiting for the CRDs to become healthy
-	// or deleted.
-	TimeoutWaitForCRD = 15 * time.Second
-	// Until is an alias for retry.Until. Exposed for tests.
-	Until = retry.Until
-)
-
 // Wait signals whether a CRD is ready or needs more time to be deployed.
 func (c *crds) Wait(ctx context.Context) error {
-	timeoutCtx, cancel := context.WithTimeout(ctx, TimeoutWaitForCRD)
-	defer cancel()
-
-	return retry.Until(timeoutCtx, IntervalWaitForCRD, func(ctx context.Context) (done bool, err error) {
-		for _, resource := range resources {
-			r := resource
-			crd := &v1.CustomResourceDefinition{}
-
-			obj, err := kubernetes.NewManifestReader([]byte(r)).Read()
-			if err != nil {
-				return retry.SevereError(err)
-			}
-
-			if err := c.client.Get(ctx, client.ObjectKeyFromObject(obj), crd); client.IgnoreNotFound(err) != nil {
-				return retry.SevereError(err)
-			}
-
-			if err := health.CheckCustomResourceDefinition(crd); err != nil {
-				return retry.MinorError(err)
-			}
-		}
-		return retry.Ok()
-	})
+	return kubernetesutils.WaitUntilCRDManifestsReady(ctx, c.client, resources)
 }
 
 // WaitCleanup for destruction to finish and component to be fully removed. crdDeployer does not need to wait for cleanup.
